requirements.txt
python-dotenv
openai
robin_stocks
pyotp
pandas
ta
requests
